requirements.txt
attrs==20.3.0
cloudpickle==1.6.0
gin-config==0.4.0
google==3.0.0
gym==0.18.0
gym-sokoban @ git+https://github.com/subgoal-search/gym-sokoban.git
joblib==1.0.1
neptune-client==0.4.130
numpy==1.19.5
psutil==5.8.0
pycolab==1.2
pylint==2.7.4
pylint-quotes==0.2.1
pytest==6.2.3
randomdict==0.2.0
ray==0.8.5
tblib==1.7.0
tensorflow==2.4.1
tqdm==4.60.0

# Install torch, but make sure that torch+CUDA versions are compatible with GPU of Turing architecture
# see: https://discuss.pytorch.org/t/cuda-error-cublas-status-internal-error-when-calling-cublascreate-handle/114341/9
torch==1.8.1+cu111
torchvision==0.9.1+cu111
torchaudio==0.8.1

transformers==4.5.1
